Correções e ajustes.

+++ database/migrations/2017_07_03_003750_create_empresas_table.php
@@ -23,7 +23,7 @@ class CreateEmpresasTable extends Migration {
 			$table->integer('endereco_id')->unsigned();
 			$table->string('foto_perfil', 100);
 			$table->integer('pontuacao')->default(0);
-			$table->integer('avaliacao_geral')->default(0);
+			$table->double('avaliacao_geral')->default(0);
 			$table->timestamps();
 			$table->rememberToken();
 			$table->string('account_confirmation');
